Name mistake in the  test.js
At line 246, the method had an incorrect name. This must have occurred when I used Control-R on Windows to replace a method name. It seems this action changed the name of the helper function from "_testBotResponseEventMessageHelperFunction" to "_testSuccessfulGreetingMsgAfterLogin" because it matched some of the characters. That has now been fixed.